+++ src/test/index.js
@@ -4,9 +4,9 @@ const { TestDirector } = require('test-director');
 
 const tests = new TestDirector();
 
-require('./universal/GraphQL.test')(tests);
-require('./universal/graphqlFetchOptions.test')(tests);
-require('./universal/hashObject.test')(tests);
-require('./universal/useGraphQL.test')(tests);
+require('./private/graphqlFetchOptions.test')(tests);
+require('./public/GraphQL.test')(tests);
+require('./public/hashObject.test')(tests);
+require('./public/useGraphQL.test')(tests);
 
 tests.run();